opsguardian-agent/requirements.txt
# Core dependencies
requests>=2.31.0
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Google Cloud dependencies
google-cloud-aiplatform[adk]>=1.36.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0

# Development dependencies
pytest>=7.4.0
black>=23.3.0
flake8>=6.0.0
